+++ test/integration/channel.js
@@ -29,9 +29,8 @@ const wsUrl = process.env.TEST_WS_URL || 'ws://localhost:3014/channel'
 plan(BigNumber('1000e18').toString())
 
 const identityContract = `
-contract Identity =
-  type state = ()
-  entrypoint main(x : int) = x
+payable contract Identity =
+  payable entrypoint main(x : int): int = x
 `
 
 function waitForChannel (channel) {
@@ -44,7 +43,7 @@ function waitForChannel (channel) {
   )
 }
 
-describe('Channel', function () {
+describe.skip('Channel', function () {
   configure(this)
   this.timeout(120000)