+++ test/integration/contract.js
@@ -23,7 +23,7 @@ import * as R from 'ramda'
 
 const identityContract = `
 contract Identity =
-  entrypoint main(x : int) = x
+ entrypoint main(x : int) = x
 `
 const stateContract = `
 contract StateContract =
@@ -31,14 +31,24 @@ contract StateContract =
   entrypoint init(value) : state = { value = value }
   entrypoint retrieve() : string = state.value
 `
+const libContract = `
+namespace TestLib =
+  function sum(x: int, y: int) : int = x + y
+`
+const contractWithLib = `
+include "testLib"
+contract Voting =
+  entrypoint sumNumbers(x: int, y: int) : int = TestLib.sum(x, y)
+`
 const testContract = `
 namespace Test =
   function double(x: int): int = x*2
 
 
 contract Voting =
-  entrypoint test() : int = 1
+  entrypoint test : () => int
 
+include "testLib"
 contract StateContract =
   type number = int
   record state = { value: string, key: number, testOption: option(string) }
@@ -47,32 +57,32 @@ contract StateContract =
   datatype dateUnit = Year | Month | Day
   
   entrypoint init(value: string, key: int, testOption: option(string)) : state = { value = value, key = key, testOption = testOption }
-  entrypoint retrieve() : (string, int) = (state.value, state.key)
+  entrypoint retrieve() : string*int = (state.value, state.key)
 
   entrypoint intFn(a: int) : int = a
-  entrypoint stringFn(a: string) : string = a
+  payable entrypoint stringFn(a: string) : string = a
   entrypoint boolFn(a: bool) : bool = a
   entrypoint addressFn(a: address) : address = a
   entrypoint contractAddress (ct: address) : address = ct
   entrypoint accountAddress (ak: address) : address = ak
 
-  entrypoint tupleFn (a: (string, int)) : (string, int) = a
-  entrypoint tupleInTupleFn (a: ((string, string), int)) : ((string, string), int) = a
-  entrypoint tupleWithList (a: (list(int), int)) : (list(int), int) = a
+  entrypoint tupleFn (a: string*int) : string*int = a
+  entrypoint tupleInTupleFn (a: (string*string)*int) : (string*string)*int = a
+  entrypoint tupleWithList (a: list(int)*int) : list(int)*int = a
   
   entrypoint listFn(a: list(int)) : list(int) = a
   entrypoint listInListFn(a: list(list(int))) : list(list(int)) = a
   
-  entrypoint mapFn(a: map(address, (string, int))) : map(address, (string, int)) = a
-  entrypoint mapOptionFn(a: map(address, (string, option(int)))) : map(address, (string, option(int))) = a
+  entrypoint mapFn(a: map(address, string*int)) : map(address, string*int) = a
+  entrypoint mapOptionFn(a: map(address, string*option(int))) : map(address, string*option(int)) = a
   
   entrypoint getRecord() : state = state
   stateful entrypoint setRecord(s: state) = put(s)
   
   entrypoint intOption(s: option(int)) : option(int) = s
-  entrypoint listOption(s: option(list((int, string)))) : option(list((int ,string))) = s
+  entrypoint listOption(s: option(list(int*string))) : option(list(int*string)) = s
   
-  entrypoint testFn(a: list(int), b: bool) : (list(int), bool) = (a, b)
+  entrypoint testFn(a: list(int), b: bool) : list(int)*bool = (a, b)
   entrypoint approve(tx_id: int, voting_contract: Voting) : int = tx_id
   
   entrypoint hashFn(s: hash): hash = s
@@ -84,8 +94,10 @@ contract StateContract =
   entrypoint datTypeFn(s: dateUnit): dateUnit = s
 `
 
-const encodedNumberSix = 'cb_AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAaKNdnK'
-
+const encodedNumberSix = 'cb_DA6sWJo='
+const filesystem = {
+  testLib: libContract
+}
 plan('1000000000000000000000')
 
 describe('Contract', function () {
@@ -100,6 +112,8 @@ describe('Contract', function () {
   })
 
   it('precompiled bytecode can be deployed', async () => {
+    const { version, consensusProtocolVersion } = contract.getNodeInfo()
+    console.log(`Node => ${version}, consensus => ${consensusProtocolVersion}, compiler => ${contract.compilerVersion}`)
     const code = await contract.contractCompile(identityContract)
     return contract.contractDeploy(code.bytecode, identityContract).should.eventually.have.property('address')
   })
@@ -110,7 +124,7 @@ describe('Contract', function () {
   })
 
   it('deploy static compiled contract', async () => {
-    const res = await bytecode.deployStatic()
+    const res = await bytecode.deployStatic([])
     res.result.should.have.property('gasUsed')
     res.result.should.have.property('returnType')
   })
@@ -143,6 +157,8 @@ describe('Contract', function () {
     const client = await BaseAe()
     client.removeAccount('ak_2a1j2Mk9YSmC1gioUq4PWRm3bsv887MbuRVwyv4KaUGoR1eiKi')
     client.addresses().length.should.be.equal(0)
+    const address = await client.address().catch(e => false)
+    address.should.be.equal(false)
     const { result } = await client.contractCallStatic(identityContract, deployed.address, 'main', ['42'])
     result.callerId.should.be.equal(client.Ae.defaults.dryRunAccount.pub)
   })
@@ -158,7 +174,7 @@ describe('Contract', function () {
   })
 
   it('initializes contract state', async () => {
-    const data = `"Hello World!"`
+    const data = '"Hello World!"'
     return contract.contractCompile(stateContract)
       .then(bytecode => bytecode.deploy([data]))
       .then(deployed => deployed.call('retrieve'))
@@ -168,6 +184,47 @@ describe('Contract', function () {
         throw e
       })
       .should.eventually.become('Hello World!')
+  })
+  describe('Namespaces', () => {
+    let deployed
+    it('Can compiler contract with external deps', async () => {
+      const filesystem = {
+        testLib: libContract
+      }
+      const compiled = await contract.contractCompile(contractWithLib, { filesystem })
+      compiled.should.have.property('bytecode')
+    })
+    it('Throw error when try to compile contract without providing external deps', async () => {
+      try {
+        await contract.contractCompile(contractWithLib)
+      } catch (e) {
+        e.message.indexOf('Couldn\'t find include file').should.not.be.equal(-1)
+      }
+    })
+    it('Can deploy contract with external deps', async () => {
+      const filesystem = {
+        testLib: libContract
+      }
+      const compiled = await contract.contractCompile(contractWithLib, { filesystem })
+      deployed = await compiled.deploy()
+      deployed.should.have.property('address')
+
+      const deployedStatic = await compiled.deployStatic([])
+      deployedStatic.result.should.have.property('gasUsed')
+      deployedStatic.result.should.have.property('returnType')
+
+      const encodedCallData = await compiled.encodeCall('sumNumbers', ['1', '2'])
+      encodedCallData.indexOf('cb_').should.not.be.equal(-1)
+    })
+    it('Can call contract with external deps', async () => {
+      const callResult = await deployed.call('sumNumbers', ['1', '2'])
+      const decoded = await callResult.decode()
+      decoded.should.be.equal(3)
+
+      const callStaticResult = await deployed.callStatic('sumNumbers', ['1', '2'])
+      const decoded2 = await callStaticResult.decode()
+      decoded2.should.be.equal(3)
+    })
   })
 
   describe('Sophia Compiler', function () {
@@ -190,7 +247,7 @@ describe('Contract', function () {
       isString.should.be.equal(true)
     })
     it('decode call-data', async () => {
-      return contract.contractDecodeCallResultAPI(identityContract, 'main', encodedNumberSix, 'ok').should.eventually.become(6)
+      return contract.contractDecodeCallResultAPI(identityContract, 'main', encodedNumberSix, 'ok', { backend: 'fate' }).should.eventually.become(6)
     })
     it('Use invalid compiler url', async () => {
       try {
@@ -206,7 +263,7 @@ describe('Contract', function () {
     let contractObject
 
     it('Generate ACI object', async () => {
-      contractObject = await contract.getContractInstance(testContract, { opt: { amount: 10000, ttl: 10 } })
+      contractObject = await contract.getContractInstance(testContract, { filesystem, opt: { ttl: 0 } })
       contractObject.should.have.property('interface')
       contractObject.should.have.property('aci')
       contractObject.should.have.property('source')
@@ -215,7 +272,9 @@ describe('Contract', function () {
       contractObject.should.have.property('compile')
       contractObject.should.have.property('call')
       contractObject.should.have.property('deploy')
-      contractObject.options.amount.should.be.equal(10000)
+      contractObject.options.ttl.should.be.equal(0)
+      contractObject.options.should.have.property('filesystem')
+      contractObject.options.filesystem.should.have.property('testLib')
       const functionsFromACI = contractObject.aci.functions.map(({ name }) => name)
       const methods = Object.keys(contractObject.methods)
       R.equals(methods, functionsFromACI).should.be.equal(true)
@@ -239,11 +298,34 @@ describe('Contract', function () {
       result.callerId.should.be.equal(onAccount)
     })
 
+    it('Can deploy using AEVM', async () => {
+      await contractObject.compile({ backend: 'aevm' })
+      const deployStatic = await contractObject.methods.init.get('123', 1, 'hahahaha', { backend: 'aevm' })
+      deployStatic.should.be.an('object')
+      deployed = await contractObject.methods.init('123', 1, 'hahahaha', { backend: 'aevm' })
+      deployed.should.be.an('object')
+      await contractObject.compile()
+    })
+
     it('Deploy contract before compile', async () => {
       contractObject.compiled = null
       await contractObject.methods.init('123', 1, 'hahahaha')
       const isCompiled = contractObject.compiled.length && contractObject.compiled.slice(0, 3) === 'cb_'
       isCompiled.should.be.equal(true)
+    })
+    it('Fail on paying to not payable function', async () => {
+      const amount = 100
+      try {
+        await contractObject.methods.intFn.send(1, { amount })
+      } catch (e) {
+        e.message.should.be.equal(`You try to pay "${amount}" to function "intFn" which is not payable. Only payable function can accept tokens`)
+      }
+    })
+    it('Can pay to payable function', async () => {
+      const contractBalance = await contract.balance(contractObject.deployInfo.address)
+      await contractObject.methods.stringFn.send('1', { amount: 100 })
+      const balanceAfter = await contract.balance(contractObject.deployInfo.address)
+      balanceAfter.should.be.equal(`${+contractBalance + 100}`)
     })
     it('Call contract on specific account', async () => {
       const current = await contract.address()
@@ -599,14 +681,14 @@ describe('Contract', function () {
         return res.decode().should.eventually.become([1, 2])
       })
       it('Call contract using using js type arguments', async () => {
-        const res = await contractObject.methods.listFn([ 1, 2 ])
+        const res = await contractObject.methods.listFn([1, 2])
         return res.decode().should.eventually.become([1, 2])
       })
       it('Call contract using using js type arguments and skip result transform', async () => {
         contractObject.setOptions({ skipTransformDecoded: true })
-        const res = await contractObject.methods.listFn([ 1, 2 ])
+        const res = await contractObject.methods.listFn([1, 2])
         const decoded = await res.decode()
-        const decodedJSON = JSON.stringify([ 1, 2 ])
+        const decodedJSON = JSON.stringify([1, 2])
         contractObject.setOptions({ skipTransformDecoded: false })
         JSON.stringify(decoded).should.be.equal(decodedJSON)
       })